logeme_backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0
psycopg2-binary>=2.9,<3.0
Pillow>=10.0,<11.0
django-storages>=1.14,<2.0
bcrypt>=4.1,<5.0
python-dotenv>=1.0,<2.0
